pyproject.toml
[project]
name = "contributor"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "sudoskys", email = "coldlando@hotmail.com"},
]
dependencies = [
    "python-dotenv>=1.0.1",
    "pydantic-settings>=2.3.3",
    "fastapi>=0.111.0",
    "loguru>=0.7.2",
    "docstring-parser>=0.16",
    "json-repair>=0.25.0",
    "tenacity>=8.4.1",
    "curl-cffi>=0.6.4",
    "Pillow>=10.3.0",
    "PyGithub>=2.3.0",
    "toml>=0.10.2",
    "aiohttp>=3.9.5",
    "requests-cache>=1.2.1",
    "pymongo>=4.8.0",
    "beanie>=1.26.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
